pyproject.toml
[tool.poetry]
name = "python-datalineage"
version = "0.1.0"
description = ""
authors = ["satokiyo <satokiyo@loop8.biz>"]
readme = "README.md"
packages = [{include = "python_datalineage"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
pip = "^23.0.1"
wheel = "^0.40.0"
setuptools = "^67.6.1"
google-cloud-datacatalog = "^3.11.1"
google-cloud-datacatalog-lineage = "^0.2.1"
google-cloud-bigquery = "^3.9.0"
pandas = "^2.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.1.1"
black = "^23.3.0"
isort = "^5.12.0"
tox = "^4.5.1"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
import_heading_stdlib      = "Standard Library"
import_heading_thirdparty  = "Third Party Library"
import_heading_firstparty  = "First Party Library"
import_heading_localfolder = "Local Library"
multi_line_output = 3
include_trailing_comma = true
force_sort_within_sections = true

[tool.flake8]
max-line-length = 88
max-complexity = 18
ignore = "E203,E266,W503,"

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = ["src/"]
addopts = "-rsxX -s -l --tb=short --strict-markers -v --ff --pdb -m test_target" # --setup-show"
markers = ["test_target: test target only"]
testpaths = [
    "tests/"
]

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py38,py39,py310
isolated_build=true
skipdist = true

[testenv]
whitelist_externals = poetry
deps=pytest
setenv = 
    PYTHONPATH = src/
commands = 
    poetry install -v --no-root
    poetry run pytest
"""


[tool.mypy]
no_strict_optional = true
show_error_context = false
show_column_numbers = true
ignore_missing_imports = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
implicit_reexport = true
namespace_packages = true
show_error_codes = true
strict_equality = true
warn_unreachable = true
warn_unused_configs = true
exclude = [
    "^test/",
    "^build/",
    "^.venv/",
    "__init__.py$",
]

[[tool.mypy.overrides]]
module = "setuptools.*"
